backend/requirements.txt
aiofiles==24.1.0
aiogram==2.23.1
aiohappyeyeballs==2.4.3
aiohttp<3.9.0,>=3.8.0
aioredis==1.3.1
aiosignal==1.3.1
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.3
asyncpg==0.30.0
attrs==24.2.0
Babel==2.9.1
bcrypt==4.1.2
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crypto==1.4.1
cryptography==43.0.3
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.7.0
email_validator==2.1.2
eval_type_backport==0.2.0
fastapi==0.115.2
fastapi-cache==0.1.0
fastapi-users==13.0.0
fastapi-users-db-sqlalchemy==6.0.1
filelock==3.16.1
frozenlist==1.4.1
greenlet==3.1.1
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jiter==0.7.0
jose==1.0.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.15
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langsmith==0.1.139
loguru==0.7.2
magic-filter==1.0.12
makefun==1.15.6
Mako==1.3.5
markdown-it-py==3.0.0
marshmallow==3.23.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
Naked==0.1.32
openai==1.53.0
orjson==3.10.11
packaging==24.1
passlib==1.7.4
prompt_toolkit==3.0.48
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pwdlib==0.2.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.8.0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
redis==5.1.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
shellescape==3.8.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.40.0
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.66.6
typer==0.12.5
typing-inspect==0.9.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.15.5
asgiref==3.8.1
Django==4.2.16
efficientnet-pytorch==0.7.1
fsspec==2024.10.0
image==1.5.33
joblib==1.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.1
threadpoolctl==3.5.0
torch==2.2.2
torchvision==0.17.2
typing_extensions==4.12.2
tzdata==2024.2
langchain_community
segmentation-models-pytorch==0.3.4
matplotlib==3.9.2